[Tests]: Use helper function for all tests in polyfill dir
