+++ src/host/polyfills/runtime.test.ts
@@ -1,28 +1,19 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { getTextFromFile } from "../../test/helpers";
+import { testPatch } from "../../test/helpers";
+import applyRuntimeImportScriptPathPrefixPatch from "./runtime";
 
 describe("importScriptPathPrefix replacement", () => {
     it("verifies that importScriptPathPrefix is exposed as a module variable", async () => {
         const filePath = "root/root.js";
-        const fileContents = getTextFromFile(filePath);
-        if (!fileContents) {
-            throw new Error(`Could not find file: ${filePath}`);
-        }
+        const expectedStrings = ["self.importScriptPathPrefix=baseUrl.substring(0,baseUrl.lastIndexOf('/')+1);})();"];
+        const unexpectedStrings = ["let importScriptPathPrefix"]
 
-        const apply = await import("./runtime");
-        const result = apply.default(fileContents);
-        const importScriptReplacementPattern =
-            "self.importScriptPathPrefix=baseUrl.substring(0,baseUrl.lastIndexOf('/')+1);})();";
-        const variableDeclarationPattern = "let importScriptPathPrefix";
-
-        expect(result).toEqual(expect.stringContaining(importScriptReplacementPattern));
-        expect(result).not.toEqual(expect.stringContaining(variableDeclarationPattern));
+        testPatch(filePath, applyRuntimeImportScriptPathPrefixPatch, expectedStrings, unexpectedStrings);
     });
 
     it("verifies that importScriptPathPrefix correctly reports a wrong replacement", async () => {
-        const apply = await import("./runtime");
-        const result = apply.default("");
+        const result = applyRuntimeImportScriptPathPrefixPatch("");
         expect(result).toEqual(null);
     });
 });
